ANDROID: mnt: Fix next_descendent
next_descendent did not properly handle the case
where the initial mount had no slaves. In this case,
we would look for the next slave, but since don't
have a master, the check for wrapping around to the
start of the list will always fail. Instead, we check
for this case, and ensure that we end the iteration
when we come back to the root.

Change-Id: I0c5fcaa428f3072776db9cafdb5513c0a0b68f7e
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 62094374